LayoutTests/fast/dom/Geolocation/cached-position-iframe-expected.txt

Tests that a cached position can be obtained in one frame after another frame has received a fresh position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Received cached position lat: 51.478, long: -0.166
PASS success is true
PASS successfullyParsed is true

TEST COMPLETE
